Extract OTelInspector
