+++ duetector/collectors/otel.py
@@ -22,31 +22,8 @@ from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExport
 from duetector.collectors.base import Collector
 from duetector.collectors.models import Tracking
 from duetector.extension.collector import hookimpl
+from duetector.otel import OTelInspector
 from duetector.utils import Singleton, get_grpc_cred_from_path
-
-
-class OTelInspector:
-    service_prefix = "duetector"
-    service_sep = "-"
-
-    @classmethod
-    def generate_service_name(cls, identifier: str) -> str:
-        return cls.service_sep.join([f"{cls.service_prefix}", f"{identifier}"])
-
-    @classmethod
-    def get_identifier(cls, service_name: str) -> Optional[str]:
-        if not service_name.startswith(cls.service_prefix):
-            return None
-
-        return service_name.replace(cls.service_prefix + cls.service_sep, "")
-
-    @classmethod
-    def generate_span_name(cls, t: Tracking) -> str:
-        return t.tracer
-
-    @classmethod
-    def get_tracer_name(cls, span_name: str) -> str:
-        return span_name
 
 
 class OTelInitiator(metaclass=Singleton):
